WW-5278 Collect common code into AbstractActionValidatorManager
